An easy way to implement custom models enums
